Move convertToFraction fn
Move convertToFraction function for better readability and infrastucture.
Add javadoc.
